# Refactor positions and spans