--- a/asciifile/src/file.rs
+++ b/asciifile/src/file.rs
@@ -11,126 +11,19 @@ pub struct AsciiFile<'m> {
 #[derive(Debug, Fail)]
 pub enum EncodingError {
     #[fail(
-        display = "input contains non-ascii character at byte offset {}: {}<?>",
-        position,
-        prev
+        display = "input contains non-ascii character at byte offset {}.",
+        position
     )]
-    NotAscii { position: usize, prev: String },
-}
-
-const ENCODING_ERROR_MAX_CONTEXT_LENGTH: usize = 80;
-
-#[derive(Debug, Eq, PartialEq, Copy, Clone)]
-pub enum LineTruncation {
-    Truncated,
-    NotTruncated,
+    NotAscii { position: usize },
 }
 
 impl<'m> AsciiFile<'m> {
-    /// return the start index of the line without the newline character (\n).
-    /// This means you can make slices that do NOT include the newline
-    /// character by simply using the return value of this function as
-    /// lower bound X in the range X..Y.
-    ///
-    /// If the cursor/byte offset is on a newline character, the newline
-    /// character will belong to the previous line, meaning the return
-    /// value will be the given byte offset and
-    /// not the byte offset of the next new line.
-    // TODO: remove these index based methods
-    pub fn get_line_end_idx(
-        &'m self,
-        byte_offset: usize,
-        max_context_length: usize,
-    ) -> (LineTruncation, usize) {
-        AsciiFile::get_line_end_idx_byteslice(self.mapping, byte_offset, max_context_length)
-    }
-
-    pub fn get_line_end_idx_byteslice(
-        mapping: &[u8],
-        byte_offset: usize,
-        max_context_length: usize,
-    ) -> (LineTruncation, usize) {
-        debug_assert!(byte_offset <= mapping.len());
-
-        let region_max_end = byte_offset + max_context_length;
-        let (truncation, region_end) = if mapping.len() > region_max_end {
-            (LineTruncation::Truncated, region_max_end)
-        } else {
-            (LineTruncation::NotTruncated, mapping.len())
-        };
-
-        let region = &mapping[byte_offset..region_end];
-
-        let newline_position = region
-            .iter()
-            .position(|&chr| chr as char == '\n')
-            .map(|pos| pos + byte_offset);
-
-        match newline_position {
-            Some(position) => (LineTruncation::NotTruncated, position),
-            None => (truncation, region_end),
-        }
-    }
-
-    /// return the end index of the line includes the newline character (\n).
-    /// This means you can make slices that do NOT include the newline
-    /// character by simply using the return value of this function as
-    /// upper bound Y in the range X..Y.
-    ///
-    /// If the cursor/byte offset is on a newline character, the newline
-    /// character will belong to the previous line, meaning the return
-    /// value will not be the given by the byte
-    /// offset, but the byte offset of the previous new line.
-    pub fn get_line_start_idx(
-        &'m self,
-        byte_offset: usize,
-        max_context_length: usize,
-    ) -> (LineTruncation, usize) {
-        AsciiFile::get_line_start_idx_byteslice(self.mapping, byte_offset, max_context_length)
-    }
-
-    fn get_line_start_idx_byteslice(
-        mapping: &[u8],
-        byte_offset: usize,
-        max_context_length: usize,
-    ) -> (LineTruncation, usize) {
-        debug_assert!(byte_offset <= mapping.len());
-
-        let (truncation, region_start) = byte_offset
-            .checked_sub(max_context_length)
-            .map(|start| (LineTruncation::Truncated, start))
-            .unwrap_or((LineTruncation::NotTruncated, 0));
-
-        let region = &mapping[region_start..byte_offset];
-
-        region
-            .iter()
-            .rposition(|&chr| chr as char == '\n')
-            .map(|pos| (LineTruncation::NotTruncated, pos + region_start + 1))
-            .unwrap_or((truncation, region_start))
-    }
-
     // cost: O(fileLen) since we need to check if all chars are ASCII
     pub fn new(mapping: &'m [u8]) -> Result<AsciiFile<'m>, EncodingError> {
+        // TODO: move validation out of constructor
+        // TODO: add nice error back
         if let Some(position) = mapping.iter().position(|c| !c.is_ascii()) {
-            let (truncation, start_idx) = AsciiFile::get_line_start_idx_byteslice(
-                mapping,
-                position,
-                ENCODING_ERROR_MAX_CONTEXT_LENGTH,
-            );
-            // We know everything until now has been ASCII
-            let prev: &str =
-                unsafe { std::str::from_utf8_unchecked(&mapping[start_idx..position]) };
-            let prev = format!(
-                "{dots}{context}",
-                context = prev,
-                dots = if truncation == LineTruncation::Truncated {
-                    "..."
-                } else {
-                    ""
-                }
-            );
-            return Err(EncodingError::NotAscii { position, prev });
+            return Err(EncodingError::NotAscii { position });
         }
 
         Ok(AsciiFile { mapping })
@@ -173,9 +66,8 @@ mod tests {
         assert!(file.is_err());
         let e = file.err().unwrap();
         println!("{:?}", e);
-        let EncodingError::NotAscii { position, prev } = e;
+        let EncodingError::NotAscii { position } = e;
         assert_eq!(position, 3);
-        assert_eq!(prev, "one");
     }
 
     #[test]
@@ -192,8 +84,7 @@ mod tests {
         let file = AsciiFile::new(&input);
         assert!(file.is_err());
 
-        let EncodingError::NotAscii { position, prev } = file.err().unwrap();
+        let EncodingError::NotAscii { position } = file.err().unwrap();
         assert_eq!(position, 0);
-        assert_eq!(prev, "");
     }
 }