--- a/asciifile/src/iter.rs
+++ b/asciifile/src/iter.rs
@@ -32,7 +32,7 @@ impl<'t> PositionIterator<'t> {
 
     /// Look at the next `n` characters without advancing the iterator.
     /// If there are less than `n` characters, return `None`.
-    pub fn peek_exactly(&self, n: usize) -> Option<Span> {
+    pub fn peek_exactly(&self, n: usize) -> Option<Span<'t>> {
         match self.peek_at_most(n) {
             None => None,
             Some(ref span) if span.as_str().len() < n => None,
@@ -60,6 +60,13 @@ impl<'t> PositionIterator<'t> {
                 let span_end = self.clone().take(n).last().unwrap();
                 Some(Span::new(span_start, span_end))
             }
+        }
+    }
+
+    pub fn matches(&self, wanted: &str) -> bool {
+        match self.peek_exactly(wanted.len()) {
+            None => wanted == "",
+            Some(span) => span.as_str() == wanted,
         }
     }