--- a/asciifile/src/iter.rs
+++ b/asciifile/src/iter.rs
@@ -1,6 +1,6 @@
 use crate::{Position, Span};
 
-#[derive(Copy, Clone, Debug)]
+#[derive(Clone, Debug)]
 pub struct PositionIterator<'t> {
     /// current iterator position. Always points to the next character/position
     /// to emit
@@ -12,9 +12,10 @@ impl<'t> Iterator for PositionIterator<'t> {
     fn next(&mut self) -> Option<Position<'t>> {
         match self.position_to_emit {
             None => None,
-            Some(position) => {
+            Some(ref position) => {
+                let position_to_emit = position.clone();
                 self.position_to_emit = position.next();
-                Some(position)
+                Some(position_to_emit)
             }
         }
     }
@@ -47,12 +48,12 @@ impl<'t> PositionIterator<'t> {
         debug_assert!(n >= 1);
         match self.position_to_emit {
             None => None,
-            Some(span_start) => {
+            Some(ref span_start) => {
                 // unwrap is save, since we are in a branch asserting
                 // that the iterator is not finished => has at least character
                 // `span_start` remaining
-                let span_end = self.take(n).last().unwrap();
-                Some(Span::new(span_start, span_end))
+                let span_end = self.clone().take(n).last().unwrap();
+                Some(Span::new(span_start.clone(), span_end))
             }
         }
     }
@@ -65,15 +66,15 @@ impl<'t> PositionIterator<'t> {
     }
 
     pub fn peek(&self) -> Option<Position<'t>> {
-        self.position_to_emit
+        self.position_to_emit.clone()
     }
 
     pub fn eof_reached(&self) -> bool {
-        self.peek().is_none()
+        self.position_to_emit.is_none()
     }
 }
 
-#[derive(Copy, Clone, Debug)]
+#[derive(Clone, Debug)]
 pub struct ReversePositionIterator<'t> {
     /// current iterator position. Always points to the next character/position
     /// to emit
@@ -89,9 +90,10 @@ impl<'t> Iterator for ReversePositionIterator<'t> {
     fn next(&mut self) -> Option<Position<'t>> {
         match self.position_to_emit {
             None => None,
-            Some(position) => {
+            Some(ref position) => {
+                let position_to_emit = position.clone();
                 self.position_to_emit = position.prev();
-                Some(position)
+                Some(position_to_emit)
             }
         }
     }