--- a/integration-tests/parser/class_no_block.invalid.mj.stderr
+++ b/integration-tests/parser/class_no_block.invalid.mj.stderr
@@ -1,0 +1,3 @@
+error: unexpected end of file
+
+Compilation aborted due to an error