Makes all constructors for spans and positions private. New positions and spans have to be generated by combination of safe conversions.

Also see notes on commit `remove EOF token` and `[refactor] position,span,asciifile to own files; remove PositionedChar`.