@reiner-dolp 2018-11-09
##TODO
- [ ] Check performance implications, I suspect a slow down.
- [ ] Reimplement Diagnostics output.