@reiner-dolp 2018-11-09
I haven't optimized anything yet.

This branch:
```
rd ~/r/comprakt> perf stat -- ./target/release/comprakt --lextest mjtest-tests/lexer/huge_file.mj > /dev/null

 Performance counter stats for './target/release/comprakt --lextest mjtest-tests/lexer/huge_file.mj':

       8066,354664      task-clock:u (msec)       #    1,000 CPUs utilized          
                 0      context-switches:u        #    0,000 K/sec                  
                 0      cpu-migrations:u          #    0,000 K/sec                  
             1.129      page-faults:u             #    0,140 K/sec                  
    14.128.593.069      cycles:u                  #    1,752 GHz                    
    34.539.193.258      instructions:u            #    2,44  insn per cycle         
     5.387.559.462      branches:u                #  667,905 M/sec                  
         8.260.726      branch-misses:u           #    0,15% of all branches        

       8,066785787 seconds time elapsed

       6,003629000 seconds user
       2,053508000 seconds sys
```
Current master:
```
rd ~/r/comprakt> perf stat -- ./target/release/comprakt --lextest mjtest-tests/lexer/huge_file.mj > /dev/null

 Performance counter stats for './target/release/comprakt --lextest mjtest-tests/lexer/huge_file.mj':

      11364,823887      task-clock:u (msec)       #    0,999 CPUs utilized          
                 0      context-switches:u        #    0,000 K/sec                  
                 0      cpu-migrations:u          #    0,000 K/sec                  
             1.130      page-faults:u             #    0,099 K/sec                  
    22.854.264.562      cycles:u                  #    2,011 GHz                    
    68.979.552.843      instructions:u            #    3,02  insn per cycle         
    12.532.006.443      branches:u                # 1102,701 M/sec                  
        17.411.736      branch-misses:u           #    0,14% of all branches        

      11,380485004 seconds time elapsed

       9,266102000 seconds user
       2,075789000 seconds sys
```