@reiner-dolp 2018-11-09
Yes! It's a really nice side effect of using spans. As the comments in the code say. Span is just a "fat" version of a slice with a 2D column-row position instead of just a byte offset. Same goes for Position, which is just a 2D index instead of a "normal array offset".

But I think there is a lot to optimize. Haven't looked into performance yet.