@reiner-dolp 2018-11-10
Empty spans not being representable is by design. Otherwise a lot of functions would be wrapped in Option<> just to deal with an empty file. e.g. the obvious behaviour being position.byte(), but also more weird behaviour like span.start_position().chr() would have to behave differently than span.get_first_chr()...