@reiner-dolp 2018-11-10
The easiest fix would be to make parse()/parse_program() return a `ParserResult<'f, ParsedProgram>` value of
```
enum ParsedProgram<'f> {
    EmptyFile,
    NonEmptyFile(ast::Program<'f>)
}
```
Just filter it directly in parse(). The semantic analysis phase can reject the empty file with a "missing main" immediatly when it recieves the `ParsedProgram`.