@joshuabach 2018-11-11
Hm. In my heart, I believe there exists a design that uses exclusive-spans, a `Spanned<EOF>` token and a never-ending iterator that works without special cases. But maybe that's an illusion... I'll do it your way