@joshuabach 2018-11-11
Done. Once all tests pass maybe you should take a look at it (you know the span/pos code better than me), otherwise I think we can merge this