@problame 2018-11-11
I checked the memory footprint and runtime impact of this change using `perf stat`: nothing significant, but positive improvements in runtime, less total number of branches, etc.